Fix Dapper DateOnly mapping and DashboardDto int/long mismatch
- Add DateOnlyTypeHandler: MySQL returns DateTime, Dapper couldn't
  cast to DateOnly? on Acteur/Realisateur models without this handler
- Change DashboardDto counts from int to long: COUNT(*) returns Int64
  from MySQL, causing materialization failure

+++ backend/Kinepalisse.Api/Services/DashboardService.cs
@@ -8,7 +8,7 @@ public class DashboardService
     private readonly DbConnectionFactory _db;
     public DashboardService(DbConnectionFactory db) => _db = db;
 
-    public record DashboardDto(int NbFilms, int NbSeancesAVenir, int NbReservationsJour, decimal TauxRemplissage);
+    public record DashboardDto(long NbFilms, long NbSeancesAVenir, long NbReservationsJour, decimal TauxRemplissage);
 
     public async Task<DashboardDto> GetAsync()
     {
